Edit readme and add character counter method
Added Character count method for future ranking improvement

+++ src/result.cpp
@@ -29,7 +29,7 @@ int Result::getResultCount(){
 	return searchResult.size();
 };
 
-int Result::getResultScore(int i){
+double Result::getResultScore(int i){
 	return searchResult[i].recordScore;
 };
